fix(brett): trim ws-handler + extract ground-objects toolbar under S1 600-line gate [T000470]
