+++ brett/src/client/board-boot.ts
@@ -25,6 +25,7 @@ import * as persons from './ui/persons';
 import * as povCamera from './pov-camera';
 import * as freeFly from './free-fly-camera';
 import * as exportUi from './ui/export';
+import * as groundObjects from './ground-objects';
 
 export async function bootBoard(): Promise<void> {
   // ── Scene ──────────────────────────────────────────────────────────
@@ -131,111 +132,6 @@ export async function bootBoard(): Promise<void> {
     pointerEvents: 'none',
   });
   document.body.appendChild(freezeBanner);
-  // ── T000468: Admin-Toolbar für Anker & Zonen (DARK-LAUNCH) ──────────────────
-  if ((window as any).__brettFeatures?.['t000468-ground-anchors']) {
-    // Only show for admins (isAdmin was resolved during Auth fetch above)
-    const isAdmin = (window as any).__brettCurrentUserIsAdmin === true;
-    if (isAdmin) {
-      const toolbar = document.createElement('div');
-      toolbar.id = 'ground-objects-toolbar';
-      Object.assign(toolbar.style, {
-        position: 'absolute',
-        bottom: '96px',
-        right: '12px',
-        display: 'flex',
-        flexDirection: 'column',
-        gap: '6px',
-        zIndex: '20',
-      });
-
-      // Anker-Button
-      const anchorBtn = document.createElement('button');
-      anchorBtn.textContent = '⚓ Anker';
-      anchorBtn.title = 'Boden-Anker setzen (Klick auf Boden)';
-      Object.assign(anchorBtn.style, {
-        fontFamily: 'var(--brett-font-mono, monospace)',
-        fontSize: '10px',
-        padding: '6px 10px',
-        background: 'rgba(200,169,110,0.15)',
-        border: '1px solid rgba(200,169,110,0.4)',
-        color: 'var(--brett-brass, #c8a96e)',
-        borderRadius: 'var(--brett-radius-sm, 8px)',
-        cursor: 'pointer',
-        textTransform: 'uppercase',
-        letterSpacing: '0.06em',
-      });
-
-      let anchorPlacingMode = false;
-      anchorBtn.addEventListener('click', () => {
-        anchorPlacingMode = !anchorPlacingMode;
-        anchorBtn.style.background = anchorPlacingMode
-          ? 'rgba(200,169,110,0.35)' : 'rgba(200,169,110,0.15)';
-        anchorBtn.title = anchorPlacingMode
-          ? 'Klicke auf den Boden, um einen Anker zu setzen (Esc abbrechen)'
-          : 'Boden-Anker setzen';
-        (window as any).__brettAnchorPlacing = anchorPlacingMode;
-      });
-
-      document.addEventListener('brett:anchor-placed', () => {
-        anchorPlacingMode = false;
-        anchorBtn.style.background = 'rgba(200,169,110,0.15)';
-        anchorBtn.title = 'Boden-Anker setzen';
-      });
-
-      // Zonen-Button
-      const zoneBtn = document.createElement('button');
-      zoneBtn.textContent = '▭ Zone';
-      zoneBtn.title = 'Bodenzone zeichnen';
-      Object.assign(zoneBtn.style, {
-        fontFamily: 'var(--brett-font-mono, monospace)',
-        fontSize: '10px',
-        padding: '6px 10px',
-        background: 'rgba(78,161,255,0.15)',
-        border: '1px solid rgba(78,161,255,0.4)',
-        color: 'var(--brett-blue, #4ea1ff)',
-        borderRadius: 'var(--brett-radius-sm, 8px)',
-        cursor: 'pointer',
-        textTransform: 'uppercase',
-        letterSpacing: '0.06em',
-      });
-
-      let zonePlacingMode = false;
-      zoneBtn.addEventListener('click', () => {
-        zonePlacingMode = !zonePlacingMode;
-        zoneBtn.style.background = zonePlacingMode
-          ? 'rgba(78,161,255,0.35)' : 'rgba(78,161,255,0.15)';
-        (window as any).__brettZonePlacing = zonePlacingMode;
-      });
-
-      toolbar.appendChild(anchorBtn);
-      toolbar.appendChild(zoneBtn);
-      document.body.appendChild(toolbar);
-
-      // Floor-click for anchor placement (wired into existing click handler)
-      renderer.domElement.addEventListener('click', (e) => {
-        if (!(window as any).__brettAnchorPlacing) return;
-        const { floor } = sceneApi as any;
-        if (!floor) return;
-        mannequin.setNdc(e);
-        const { ndc } = mannequin.getTickRefs();
-        if (!ndc) return;
-        raycaster.setFromCamera(ndc, camera);
-        const hits = raycaster.intersectObject(floor);
-        if (hits.length > 0) {
-          const pt = hits[0].point;
-          const ws = getWs();
-          if (isWsReady() && ws) {
-            ws.send(JSON.stringify({
-              type: 'anchor_create',
-              anchor: { x: Math.round(pt.x * 10) / 10, z: Math.round(pt.z * 10) / 10 },
-            }));
-          }
-          (window as any).__brettAnchorPlacing = false;
-          document.dispatchEvent(new CustomEvent('brett:anchor-placed'));
-        }
-      }, { capture: true });
-    }
-  }
 
   // ── T000470: Undo/Redo-Buttons (isAdmin-only, Dark-Launch) ────────────────
   const undoBtn = document.createElement('button');
@@ -333,6 +229,11 @@ export async function bootBoard(): Promise<void> {
 
   // ── Drag handling ──────────────────────────────────────────────────
   const { raycaster } = mannequin.getTickRefs();
+
+  // ── T000468: Admin-Toolbar für Anker & Zonen (DARK-LAUNCH) ──────────────────
+  if ((window as any).__brettFeatures?.['t000468-ground-anchors']) {
+    groundObjects.initGroundObjectsToolbar(renderer, sceneApi, camera, raycaster, mannequin);
+  }
 
   renderer.domElement.addEventListener('mousedown', (e) => {
     if (ui.placingMode && e.button === 0) {
